+++ app/Http/Controllers/TagsController.php
@@ -5,9 +5,13 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Tags;
+use App\Http\Traits\HasAuthorization;
+use App\Models\User;
 
 class TagsController extends Controller
 {
+
+    use HasAuthorization;
     /**
      * Display a listing of the resource.
      *
@@ -15,11 +19,14 @@ class TagsController extends Controller
      */
     public function index()
     {
+        if (!$this->isAuthorized("userOrAdmin", User::class)) {
+            return response()->json([
+                "message" => "User has not the right permissions."
+            ], 401);
+        }
         $tags = Tags::all();
         return $tags;
     }
-
-
 
     /**
      * Store a newly created resource in storage.
@@ -29,6 +36,12 @@ class TagsController extends Controller
      */
     public function store(Request $request)
     {
+        if (!$this->isAuthorized("userOrAdmin", User::class)) {
+            return response()->json([
+                "message" => "User has not the right permissions."
+            ], 401);
+        }
+
         $tags = new Tags();
         $tags->recipe_id = $request->recipe_id;
         $tags->name = $request->name;
@@ -43,6 +56,12 @@ class TagsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!$this->isAuthorized("userOrAdmin", User::class)) {
+            return response()->json([
+                "message" => "User has not the right permissions."
+            ], 401);
+        }
+
         $tags = Tags::findOrFail($id);
         $tags->recipe_id = $request->recipe_id;
         $tags->name = $request->name;
@@ -57,6 +76,12 @@ class TagsController extends Controller
      */
     public function destroy($id)
     {
+        if (!$this->isAuthorized("userOrAdmin", User::class)) {
+            return response()->json([
+                "message" => "User has not the right permissions."
+            ], 401);
+        }
+
         Tags::destroy($id);
     }
 }
