+++ routes/api.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\RecipeController;
 use App\Http\Controllers\RoleController;
 use App\Http\Controllers\PermissionController;
 use App\Http\Controllers\UserController;
+use App\Http\Controllers\TagsController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -26,7 +27,7 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 });
 
 /* ***************************************************************** Admin routes ***************************************************************** */
-Route::get("/v1/users", [AuthController::class, "index"]);
+Route::get("/v1/users", [AuthController::class, "index"])->middleware("auth:sanctum")->middleware("auth:sanctum");
 Route::post("/v1/register/admin", [AuthController::class, 'registerAdmin'])->middleware("auth:sanctum");
 Route::post("/v1/register/user", [AuthController::class, 'registerUser']);
 Route::post("/v1/login", [AuthController::class, 'login']);
@@ -36,24 +37,24 @@ Route::delete("/v1/user/{user}", [AuthController::class, "destroy"])->middleware
 /* ************************************************************************************************************************************************ */
 
 /* **************************************************************** Recipes routes **************************************************************** */
-Route::get("/v1/recipes", [RecipeController::class, "index"]);
-Route::post("/v1/recipe", [RecipeController::class, "store"]);
-Route::put("/v1/recipe/{id}", [RecipeController::class, "update"]);
-Route::delete("/v1/recipe/{id}", [RecipeController::class, "destroy"]);
+Route::get("/v1/recipes", [RecipeController::class, "index"])->middleware("auth:sanctum");
+Route::post("/v1/recipe", [RecipeController::class, "store"])->middleware("auth:sanctum");
+Route::put("/v1/recipe/{id}", [RecipeController::class, "update"])->middleware("auth:sanctum");
+Route::delete("/v1/recipe/{id}", [RecipeController::class, "destroy"])->middleware("auth:sanctum");
 /* ************************************************************************************************************************************************ */
 
 /* *************************************************************** Ingredient routes ************************************************************** */
-Route::get("/v1/ingredients", [IngredientController::class, "index"]);
-Route::get("/v1/ingredient/{ingredient}", [IngredientController::class, "show"]);
-Route::post("/v1/ingredient", [IngredientController::class, "store"]);
-Route::put("/v1/ingredient/{id}", [IngredientController::class, "update"]);
-Route::delete("/v1/ingredient/{id}", [IngredientController::class, "destroy"]);
+Route::get("/v1/ingredients", [IngredientController::class, "index"])->middleware("auth:sanctum");
+Route::get("/v1/ingredient/{ingredient}", [IngredientController::class, "show"])->middleware("auth:sanctum");
+Route::post("/v1/ingredient", [IngredientController::class, "store"])->middleware("auth:sanctum");
+Route::put("/v1/ingredient/{id}", [IngredientController::class, "update"])->middleware("auth:sanctum");
+Route::delete("/v1/ingredient/{id}", [IngredientController::class, "destroy"])->middleware("auth:sanctum");
 /* ************************************************************************************************************************************************ */
 /* *************************************************************** Tags routes ************************************************************** */
-Route::get("/v1/tags", [TagsController::class, "index"]);
-Route::post("/v1/tag", [TagsController::class, "store"]);
-Route::put("/v1/tag/{id}", [TagsController::class, "update"]);
-Route::delete("/v1/tag/{id}", [TagsController::class, "destroy"]);
+Route::get("/v1/tags", [TagsController::class, "index"])->middleware("auth:sanctum");
+Route::post("/v1/tag", [TagsController::class, "store"])->middleware("auth:sanctum");
+Route::put("/v1/tag/{id}", [TagsController::class, "update"])->middleware("auth:sanctum");
+Route::delete("/v1/tag/{id}", [TagsController::class, "destroy"])->middleware("auth:sanctum");
 /* ************************************************************************************************************************************************ */
 
 /* ***************************************************************** Roles routes ***************************************************************** */
@@ -73,14 +74,14 @@ Route::delete("/v1/permission/{permission}", [PermissionController::class, "dest
 /* ************************************************************************************************************************************************ */
 
 /* *************************************************************** Categories routes ************************************************************** */
-Route::get("/v1/categories", [CategoriesController::class, "index"]);
-Route::post("/v1/category", [CategoriesController::class, "store"]);
-Route::put("/v1/category/{id}", [CategoriesController::class, "update"]);
-Route::delete("/v1/category/{id}", [CategoriesController::class, "destroy"]);
+Route::get("/v1/categories", [CategoriesController::class, "index"])->middleware("auth:sanctum");
+Route::post("/v1/category", [CategoriesController::class, "store"])->middleware("auth:sanctum");
+Route::put("/v1/category/{id}", [CategoriesController::class, "update"])->middleware("auth:sanctum");
+Route::delete("/v1/category/{id}", [CategoriesController::class, "destroy"])->middleware("auth:sanctum");
 /* ************************************************************************************************************************************************ */
 
 /* ***************************************************************** Image routes ***************************************************************** */
-Route::get("/image/user/{filename}", [ImageController::class, "userImage"]);
-Route::get("/image/recipe/{filename}", [ImageController::class, "recipeImage"]);
-Route::get("/image/ingredient/{filename}", [ImageController::class, "ingredientImage"]);
+Route::get("/image/user/{filename}", [ImageController::class, "userImage"])->middleware("auth:sanctum");
+Route::get("/image/recipe/{filename}", [ImageController::class, "recipeImage"])->middleware("auth:sanctum");
+Route::get("/image/ingredient/{filename}", [ImageController::class, "ingredientImage"])->middleware("auth:sanctum");
 /* ************************************************************************************************************************************************ */
